:test_tube: adicionando teste unitario validar se produto nao corresponde ao setor e criacao estrutura inicial de produtos
