+++ src/test/java/br/com/meli/projetointegrador/model/service/SectionServiceTest.java
@@ -1,10 +1,8 @@
 package br.com.meli.projetointegrador.model.service;
 
 import br.com.meli.projetointegrador.exception.SectionException;
-import br.com.meli.projetointegrador.model.entity.Product;
 import br.com.meli.projetointegrador.model.entity.Section;
 import br.com.meli.projetointegrador.model.entity.Warehouse;
-import br.com.meli.projetointegrador.model.repository.ProductRepository;
 import br.com.meli.projetointegrador.model.repository.SectionRepository;
 import org.junit.jupiter.api.Test;
 
@@ -16,9 +14,7 @@ import static org.mockito.Mockito.*;
 public class SectionServiceTest {
 
     private SectionRepository mockSectionRepository = mock(SectionRepository.class);
-    private ProductRepository mockProductRepository = mock(ProductRepository.class);
     private SectionService sectionService = new SectionService(mockSectionRepository);
-    private ProductService productService = new ProductService(mockProductRepository);
 
     @Test
     void validSectionExistTest(){
@@ -74,40 +70,4 @@ public class SectionServiceTest {
 
         assertTrue(expectedMessage.contains(sectionException.getMessage()));
     }
-
-    @Test
-    void validProductSectionExistTest(){
-
-        Warehouse warehouse = new Warehouse()
-                .warehouseCode("SP")
-                .warehouseName("Sao Paulo")
-                .build();
-
-        Section section = new Section()
-                .id("1")
-                .sectionCode("LA")
-                .sectionName("Laticionios")
-                .maxLength(10)
-                .warehouse(warehouse)
-                .build();
-
-        Product product = new Product()
-                .id("1")
-                .productCode("LEI")
-                .productName("Leite")
-                .section(section)
-                .build();
-
-        when(mockProductRepository.findBySection(any()))
-                .thenReturn(Optional.of(new Product()
-                        .id("1")
-                        .productCode("LEI")
-                        .productName("Leite")
-                        .section(section)
-                        .build()));
-
-        assertTrue(productService.validProductSection(product));
-    }
-
-
 }
